requirements.txt
amqp==5.2.0
asgiref==3.7.2
beautifulsoup4==4.12.3
billiard==4.2.0
bs4==0.0.2
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==4.2.10
django-celery-beat==2.5.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
idna==3.6
kombu==5.3.5
numpy==1.26.4
pandas==2.2.0
prompt-toolkit==3.0.43
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2024.1
redis==5.0.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
